\name{exptable}
\alias{exptable}
\docType{data}
\title{Sample table for the generation of a database of standards (LCMS)}
\description{An example of a table used to construct a database of
  standards. This particular table is used to obtain object \code{LCDBtest}.}
\usage{data(exptable)}
\format{
  \code{exptable} is a \code{data.frame} summarizing information on the
  chemical standards used for creating a database.
  
  Column description:
  \describe{
    \item{ChemSpiderID}{The Chemspider ID of a specific compound.}
    \item{compound}{A string containing the (human) readable name of the
      compound.} 
    \item{formula}{The formula of the compound.}
    \item{M.ref}{The theoretical value for the observed m/z reference
      ion. This can be, for example, the protonated (de-protonated) molecular
      ion, a known adduct or a characteristic fragment.}
    \item{mz.observed}{The manually validated m/z value of the reference
      ion.} 
    \item{RTman}{The manually validated retention time value for the standard.}
    \item{stdFile}{The name (including the path) of the raw data file of the
      standard.} 
  }
}

\details{
 For a successful generation of the DB the mandatory fields are: \code{ChemSpiderID, mz.observed, RTman, stdFile}

 In the current implementation the \code{M.ref} value is not used in the creation of the DB. The difference between \code{M.ref} and \code{m.observed}, however, could be used to construct the mass error surface used during feature annotation.
 
 The sample table is also an element (\code{Reftable}) of the LCDB list generated by \code{createSTDdbLC}


}

\author{Pietro Franceschi}
\examples{
if (require(metaMSdata)) {
  ##  load the manually curated table for the standards
  data(exptable)
  ## add location of cdf file from which the standards DB is going to be
  ## built - this depends on your platform and requires the metaMSdata package
  cdfpath <- system.file("CDF_LC", package = "metaMSdata")
  
  ## files 
  files <- list.files(cdfpath, "CDF", full.names=TRUE)
  
  ## get the complete names for the files
  exptable$stdFile <- sapply(exptable$stdFile, function(x){
    out <- files[grep(x,files)]
  })
  
  ## load the settings for the analysis
  data(FEMsettings)
  
  ## set the minimum number of features to 2
  FEMsettings$Synapt.QTOF.RP$DBconstruction$minfeat  <- 2
  
  ## create the DB
  LCDBtest <- createSTDdbLC(stdInfo=exptable, 
                            settings = FEMsettings$Synapt.QTOF.RP,
                            polarity = "positive",
                            Ithr = 20)
                           
  ## saved in "LCDBtest.RData" in the data directory of the metaMS
  ## package
}
}
\keyword{datasets}
