\name{GCset}
\alias{GCset}
\docType{data}
\title{GC-MS data for repeated injections of a mixture of chemical standards}
\description{
  This object is the result of a call to function \code{xcmsSet} for the
  CDF GC-MS files present in the \code{metaMSdata} package. 
}
\usage{data(GCset)}
\keyword{datasets}
