\name{generateStdDBGC}
\alias{generateStdDBGC}
\title{Convert an msp object into a GC database object}
\description{
This function is not meant to be called directly by the user - use
createSTDdbGC instead.}
\usage{
generateStdDBGC(totalXset, settings, extDB = NULL, manualDB = NULL)
}
\arguments{
  \item{totalXset}{A list of xset objects, as generated by
    \code{peakDetection}.}
  \item{settings}{Settingslist for generating the in-house DB.}
  \item{extDB}{External DB to double-check pseudospectra of standards,
    e.g. the NIST DB in msp format.}
  \item{manualDB}{list of compounds in msp format that will be included
    virtually unchanged in the in-house DB.}
}
\value{An msp object containing the in-house DB.}
\author{Ron Wehrens}
\seealso{\code{\link{createSTDdbGC}}}
\keyword{manip}
