\name{LCresults}
\alias{LCresults}
\docType{data}
\title{Results of metaMS for a small LC-MS data set}
\description{This data file contains the final results of
  applying the metaMS pipeline to a small LC-MS data set consisting of
  repeated injections of a set of chemical standards. Its aim is to
  demonstrate the structure of the outcomes without
  having to do the slightly time-consuming calculations.
}
\usage{data(LCresults)}
\details{This data objects contains the output list of \code{runLC}
  }
\keyword{datasets}
