\name{LCxset}
\alias{LCxset}
\docType{data}
\title{Results of the peak pickingfor a small LC-MS data set}
\description{This data file contains the \code{xcmsSet} object resulting from the peak picking 
of a small LC-MS dataset (the raw files are provided inside the \code{metaMSdata} package)
}
\usage{data(LCxset)}
\details{This data objects contains the \code{xcmsSet} object
  }
\keyword{datasets}
