\name{metaSetting-methods}
\docType{methods}
%\alias{metaSetting-methods}
\alias{metaSetting}
\alias{metaSetting<-}
\alias{metaSetting, metaMSsettings-method}
\alias{metaSetting<-, metaMSsettings-method}
\title{Get or set values in metaMSsettings objects}
\description{
  Accessor function for \code{metaMSsettings} objects, allowing to get
  or set values from individual slots.
}
\section{Methods}{
  \describe{
\item{\code{signature(object = "metaMSsettings")}}{
  Get or set values from individual slots in a \code{metaMSsettings} objects.
}
}}
\seealso{\code{\link{FEMsettings}}}
\keyword{methods}
