\name{GCresults}
\alias{GCresults}
\docType{data}
\title{Results of metaMS for a small GC-MS data set}
\description{This data file contains the results of
  applying the metaMS pipeline to a small GC-MS data set consisting of
  repeated injections of a set of chemical standards. Its aim is to
  demonstrate the structure of the (intermediate) outcomes without
  having to do the slightly time-consuming calculations.
}
\usage{data(GCresults)}
\details{Data object GCresults is the result of a complete execution of
  \code{runGC}, using the example database for annotation purposes.}
\seealso{\code{\link{runGC}}}
\keyword{datasets}
