\name{LCresults}
\alias{LCresults}
\alias{LCxset}
\docType{data}
\title{Results and intermediate results of metaMS for a small LC-MS data set}
\description{This data file contains intermediate and final results of
  applying the metaMS pipeline to a small LC-MS data set consisting of
  repeated injections of a set of chemical standards. Its aim is to
  demonstrate the structure of the outcomes without
  having to do the slightly time-consuming calculations.
}
\usage{data(LCresults)}
\details{Data object \code{LCresult} contains the output list of
  \code{runLC}, whereas \code{LCxset} contains the result of XCMS peak
  picking only.}
\keyword{datasets}
